+++ ios/Classes/TalsecHandlers.swift
@@ -11,6 +11,7 @@ private let missingSecureEnclaveValue = 1564314755
 private let deviceChangeValue = 1806586319
 private let deviceIDValue =  1514211414
 private let unofficialStoreValue = 629780916
+private let systemVPNValue = 659382561
 
 /// Extension with submits events to plugin
 extension SecurityThreatCenter: SecurityThreatHandler {
@@ -46,6 +47,8 @@ extension SecurityThreat {
             return deviceIDValue
         case .unofficialStore:
             return unofficialStoreValue
+        case .systemVPN:
+            return systemVPNValue
         @unknown default:
             return unknownValue
         }