+++ ios/TalsecRuntime.xcframework/ios-arm64/TalsecRuntime.framework/Headers/CurlWrapper.h
@@ -16,7 +16,7 @@
 #include <unistd.h>
 #include <string.h>
 
-struct QPRJYVLZQhnk {
+struct fCWAVoznQoPn {
     char *memory;
     size_t size;
     CURLcode ret;