+++ ios/TalsecRuntime.xcframework/ios-arm64/TalsecRuntime.framework/Headers/TalsecRuntime-Swift.h
@@ -281,7 +281,7 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 
 #if defined(__OBJC__)
 
-SWIFT_EXTERN void __XNZNoaKkutzQfbZIdSWjPsh(void);
+SWIFT_EXTERN void __ZviRgaYbEIxXcUVozJhTbVV(void);
 
 #endif
 #if defined(__cplusplus)