+++ test/test_utils/mock_method_channel.dart
@@ -23,7 +23,7 @@ class MockMethodChannel {
     this.result,
     this.delay = Duration.zero,
   }) : methodChannel = MethodChannel(channelName) {
-    TestDefaultBinaryMessengerBinding.instance!.defaultBinaryMessenger
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
         .setMockMethodCallHandler(methodChannel, _handler);
   }